attendance_system/requirements.txt
Flask==2.0.1
face-recognition==1.3.0
opencv-python==4.5.3.56
mysql-connector-python==8.0.26
python-dotenv==0.19.0
qrcode==7.3
pyzbar==0.1.8
google-auth-oauthlib==0.4.6
pillow==8.3.2
flask-mail==0.9.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
Werkzeug==2.0.1
gunicorn==20.1.0
